Introduce standalone `runGraph` method, switch `BoardRunner.run` to use it.
