+++ packages/breadboard/tests/node/run/state.ts
@@ -1,11 +1,12 @@
-import { ok } from "node:assert";
-import test, { describe } from "node:test";
-
 /**
  * @license
  * Copyright 2024 Google LLC
  * SPDX-License-Identifier: Apache-2.0
  */
+
+import { ok } from "node:assert";
+import test, { describe } from "node:test";
+
 describe("managed run state", () => {
   test("stack manager", async () => {
     ok(true);
